Refactorized routes
